chore: remove now-unused code due to new payment API
