+++ api/src/routes/payment/index.ts
@@ -1,8 +1,8 @@
 import './quotes.js'
 import './initiate.js'
 import './redirect.js'
-import './status2.js'
+import './status.js'
 
 export type { PaymentQuoteInput, PaymentQuoteResult } from './quotes.js'
 export type { PaymentInitiateInput, PaymentInitiateResult } from './initiate.js'
-export type { PaymentStatus } from './status2.js'
+export type { PaymentStatus } from './status.js'
