+++ frontend/app/routes/payment-confirmation.tsx
@@ -1,12 +1,4 @@
-import { useEffect, useRef } from 'react'
-import {
-  useLoaderData,
-  data,
-  type LoaderFunctionArgs,
-  type MetaFunction,
-} from 'react-router'
-import { KV_PAYMENTS_PREFIX } from '@shared/types'
-import { validatePaymentParams } from '~/utils/validate.server'
+import type { MetaFunction } from 'react-router'
 
 export const meta: MetaFunction = () => {
   return [
@@ -15,62 +7,8 @@ export const meta: MetaFunction = () => {
   ]
 }
 
-export async function loader({ request, context }: LoaderFunctionArgs) {
-  const { env } = context.cloudflare
-  const url = new URL(request.url)
-  const params = Object.fromEntries([...url.searchParams])
-
-  const validation = validatePaymentParams(params)
-  if (!validation.success) {
-    return data({
-      success: false,
-      error: 'Invalid parameters',
-      params,
-    })
-  }
-
-  const { paymentId } = validation.data
-
-  try {
-    const existingData = await env.PUBLISHER_TOOLS_KV.get(
-      KV_PAYMENTS_PREFIX + paymentId,
-    )
-
-    if (existingData) {
-      // avoids spamming the KV store with redundant entries for the same payment
-      return data({ success: true, message: 'Already stored', params })
-    }
-
-    await env.PUBLISHER_TOOLS_KV.put(
-      KV_PAYMENTS_PREFIX + paymentId,
-      JSON.stringify(params),
-      {
-        expirationTtl: 300, // 5min
-      },
-    )
-
-    return data({ success: true, params })
-  } catch {
-    return data(
-      { success: false, error: 'Failed to store data', params },
-      { status: 500 },
-    )
-  }
-}
-
+// This page has URL params set by /payment/redirect API
 export default function PaymentComplete() {
-  const { params } = useLoaderData<typeof loader>()
-  const hasPostedMessage = useRef(false)
-
-  useEffect(() => {
-    if (hasPostedMessage.current) return
-
-    hasPostedMessage.current = true
-    if (window.opener) {
-      window.opener.postMessage({ type: 'GRANT_INTERACTION', ...params }, '*')
-    }
-  }, [])
-
   return (
     <div className="min-h-screen flex items-center justify-center bg-gray-50">
       <div className="max-w-md w-full space-y-8 p-8">
