+++ frontend/app/utils/validate.server.ts
@@ -1,8 +1,3 @@
-import type {
-  PaymentStatus,
-  PaymentStatusRejected,
-  PaymentStatusSuccess,
-} from 'publisher-tools-api/src/routes/payment/status'
 import z from 'zod'
 import {
   checkHrefFormat,
@@ -34,23 +29,3 @@ export const walletSchema = z.object({
       }
     }),
 })
-
-export const PaymentStatusSuccessSchema = z.object({
-  paymentId: z.string().min(1, 'Payment ID is required'),
-  hash: z.string().min(1, 'Hash is required'),
-  interact_ref: z.string().min(1, 'Interact reference is required'),
-}) satisfies z.ZodType<PaymentStatusSuccess>
-
-export const PaymentStatusRejectedSchema = z.object({
-  paymentId: z.string().min(1, 'Payment ID is required'),
-  result: z.literal('grant_rejected'),
-}) satisfies z.ZodType<PaymentStatusRejected>
-
-const PaymentStatusSchema = z.union([
-  PaymentStatusSuccessSchema,
-  PaymentStatusRejectedSchema,
-]) satisfies z.ZodType<PaymentStatus>
-
-export const validatePaymentParams = (params: Record<string, string>) => {
-  return PaymentStatusSchema.safeParse(params)
-}
